refactor : 코드 리뷰 피드백 반영_pr 상세 설명 참조
